chore: generalise lemmas to `ENorm` spaces, part 1 (#21380)
Extracted from #21375: these lemmas only change the assumption on the co-domain, but not any hypothesis.

This continues the work started in #20121, #20122 and #21306: those PRs generalised the definitions resp. made lemma statements use the enorm. This PR starts generalising the lemmas using those definitions.

This work is part of (and a necessary pre-requisite for) the Carleson project.