feat(CI): bench-after-CI (#21414)
Adds a CI step that checks if the `bench-after-CI` label is present.  If so, it adds a comment `!bench` once CI passes, triggering a benchmarking run.

The management is as follow:
* the label `bench-after-CI` is applied manually while CI is running;
* once CI passes the lint phase, if the label is present, a bot comments `!bench`, starting the bench process;
* when the bench process terminates, it posts the report;
* this triggers the bench-summary bot to post the summary *and* to remove the label.

Meanwhile, the label `bench-after-CI` is in the list of labels that `bors` rejects.

[Zulip](https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/.60bench-after-CI.60.3F)

I added the label to this PR, ~~but I am not sure whether the workflow will act on it, since the code is not in `master`~~ and the automation worked.