feat(LinearAlgebra/ExteriorPower): add iMulti_family definition for product of a family of vectors (#21397)
This splits the definition of `ιMulti_family` from PR #10654 along with a few basic lemmas that do not rely on other parts of the PR. The `ιMulti_family` definition allows taking the product of a subset of a family of vectors, which is useful when the number of vectors is different from the degree of the exterior power. In particular, this is useful for working with a collection of basis or spanning vectors.