feat(AlgebraicTopology): notation X ^[n] for cosimplicial objects (#21485)
We change the notation for the `n`-th term of a cosimplicial object `X`, introducing the new notation `X ^[n]`. The previous notation `X _[n]` is also the notation used when `X` is a simplicial object. This change prevents any confusion between the two and is more in line with the established notation `Xⁿ` for cosimplicial objects (with simplicial objects denoted `Xₙ`).