feat(BigOperators/Finset): prod_comm_3
A cyclically permuting version of `Finset.prod_comm`. Sort of like `LinearMap.trace_mul_cycle`. Think you could debate if it's "necessary" but we were using it enough that we wanted to have its own lemma; the alternative would be writing `prod_comm` with implicit arguments a bunch of times.