feat(AlgebraicTopology): notations X _⦋n⦌ and X ^⦋n⦌ for (co)simplicial objects (#21581)
We change the notation for the `n`-th term of a simplicial object `X` from `X _[n]` to `X _⦋n⦌`. We also change the notation for the `n`-th term of a cosimplicial object `X` from `X ^[n]` to `X ^⦋n⦌`.

This change makes the notations for (co)simplicial objects consistent with the notation for objects of the simplex category, which was updated in #21565.