fix(Mathlib/Tactic): `minImports` now correctly handles namespaced declarations (#21505)
The `minImports` linter/tactic is passed a piece of syntax and needs to figure out the dependencies of the declaration. Apart from looking at the syntax itself, it also looks at the elaborated term. This goes via determining the `Name` associated to that piece of syntax. The previous algorithm would get confused if there's a name `_root_.foo` and we're looking at `namespace Foo; def foo`. By replicating the logic of the builtin `Lean.Elab.mkDeclName` (that function unfortunately mixes name parsing/mangling and environment updating, so we can't call into it directly).

Apart from the extra test case, I have made sure that `MeasureTheory.Measure.le_add_left` now reports a sensible set of minimal imports. I have not added it as a test case since it is so far in the import hierarchy.

This PR also updates `upstreamableDecl` to use the same syntax -> name resolution.