feat(Analysis/Normed/Affine/DistPt): distance between point and affine subspace
Add a definition `AffineSubspace.distPt` for the distance between a
point and an affine subspace (line, plane, etc.), defined as the
infimum of distances to any point in that subspace.

This is a standard geometrical notion, currently represented when
needed in the Euclidean context with `dist p (orthogonalProjection s p)`
and is commonly enough used that it seems appropriate for it to have
its own definition and API rather than requiring the more verbose form
with `orthogonalProjection` to be used.  I have not however tried to
work out which existing lemmas might most appropriately be restated
with the new definition and which are more useful in the
`orthogonalProjection` form, although I've included the lemma
`dist_orthogonalProjection_eq_distPt` equating the two in the
Euclidean context.

(One specific motivation for this definition is stating IMO 2020 P6,
but the concept is certainly used far beyond that specific problem.)