feat(Geometry/Euclidean/Sphere/Basic): diameters
Define `EuclideanGeometry.Sphere.IsDiameter`, the property of two
points being the endpoints of a diameter of a sphere, and
`EuclideanGeometry.Sphere.ofDiameter`, to construct the sphere on a
given diameter, and set up some basic API for these definitions.

This enables stating and proving geometrical results involving
diameters (as shown in the API setup here, there are several different
equivalent characterizations of a diameter, so justifying having a
common definition and API showing the different characterizations are
equivalent).  In particular, having the definition prepares the way
for stating and proving Thales's theorem, an easy entry on the
1000-plus list we're currently missing.