chore(Matroid): rename `Rk` -> `Rank` in typeclass names (#21605)
as per the (weak) consensus [on zulip](https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/Matroid.20rank.20-.20naming.20poll) and the convention established in `Data.Matroid.Rank.Cardinal`, the rank of a matroid is referred to as 'rank' in lemma names, and the rank of a set is referred to as `rk`. 

Accordingly (and for internal consistency), we rename the typeclasses `RkPos` -> `RankPos`, `FiniteRk` -> `RankFinite` and `InfiniteRk` -> `RankInfinite`.