chore(Data/Matroid/IndepAxioms): improved docstring for `IndepMatroid` (#21180)
This PR changes the docstring of `IndepMatroid` to more explicitly lay out the intended design pattern for using `IndepMatroid` to define a matroid. 

This felt like it might be helpful, because two people who spoke to me after using the definition were experiencing difficulties due to going against the pattern, so it wasn't so obvious what to do from the existing docstring. 

@alreadydone @madvorak