chore(Order/Filter): split `Filter/Bases.lean`
The goal of this PR is to move the requirement for finiteness further down the import graph. To do so, this PR splits `Filter/Bases.lean` into three files:
* `Bases/Defs.lean`: definition of `FilterBases`, `IsFilterBasis`, `Filter.HasBasis`
* `Bases/Basic.lean`: basic results (those do not depend on finiteness)
* `Bases/Finite.lean`: results that depend on finiteness

We then use this split to splits `AddTopBot/Basic.lean` further, moving out finiteness to `AtTopBot/Finite.lean` and upstreaming results on products to `AtTopBot/Prod.lean`.

This should hopefully reduce the amount of finiteness that is needed to work with topological and metric spaces.