fix(minImports): reset state after parsing attributes (#21786)
Yet another bug fix, this time found thanks to [this run](https://github.com/leanprover-community/mathlib4/actions/runs/13284754011) of the late importers report.

Here is the underlying issue.  The declaration
```lean
@[scoped simp] theorem XX.YY : 0 = 0 := rfl
```
accepts the `scoped` modifier to the `simp` attribute, *because* the namespacing is taken care of by `XX.YY`.  However, when the `minImports` machinery parses the attribute, it does not have access to the declaration name and this triggers an error if there is no explicit open namespace.