feat(CategoryTheory): the retract argument (#20666)
In a category, if `i ≫ p = f`, and `f` has the left lifting property with respect to `p`, then `f` is a retract of `i`. This statement and its dual are formalised, as well as consequences for classes of morphisms which satisfy `HasFactorization W₁ W₂`. This is applied to classes of morphisms in model categories in #20210.



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>