chore(Matroid/Dual): desimp dual_indep_iff_exists (#21687)
It was never a good idea for this to be a simp lemma, which was my bad decision in the first place - dual independence is often best spelled `Coindep`, and the RHS of this one often makes things actively more complicated. 

The tiny diff indicates this lemma was never used in simplification.