feat(SetTheory): regular cardinals have a least element (#21780)
`Ordinal.toTypeOrderBotOfPos` is deprecated in favour of a new definition `Ordinal.toTypeOrderBotOfPos` which takes an assumption `o ≠ 0`. A similar definition is introduced for cardinals, and a lemma `Cardinal.IsRegular.ne_zero` is added. As a result, if `c` is a regular cardinal, then `c.ord.toType` has a least element. 



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>