feat(Algebra/Module/Submodule/Lattice): Add explicit value of sSup for submodules (#21850)
On [zulip](https://leanprover.zulipchat.com/#narrow/channel/113489-new-members/topic/Where.20is.20sSup.20defined.20on.20submodules.3F) it was noted that `sSup` was not explicitly defined in the `CompleteLattice (Submodule R M)` instance, not even as a default value in some class. This caused confusion about the source of the definition, as after the instance, there clearly was *some* value to the field. It emerged that lean was able to fill in the field through unification of the `sSup_le` field, but it was suggested that the value of the field should be made explicit in order to prevent such confusion.