feat(Combinatorics/SimpleGraph): Characterizing `SimpleGraph.IsCycles` (#20830)
Shows that IsCycles is appropriately named. The core result is `IsCycles.exists_cycle_toSubgraph_verts_eq_connectedComponent_supp`, which shows that for an edge in a nontrivial component a cycle can be constructed that has exactly the edges of the component.

In order to get this some API is included:

- Coercions of walks in graphs that are the result of coerced subgraphs
- NeighborSets of `toSubgraph` for paths
- Supporting results for `IsCycles`
- Some miscellaneous supporting lemma's

This is all in preparation for Tutte's theorem. 



Co-authored-by: Pim Otte <otte.pim@gmail.com>