chore(CI): add awaiting-CI label when a build finishes
If the build passed, the label will be removed shortly thereafter;
for failing builds, the label will stay around (which is a small change in
behaviour).
This has the side-effect that completing CI always causes a visible state
change to a PR, so e.g. the queueboard can pick up this change.