redo order_dual tags
Try to use the reorder option as little as possible. Often this means we have to add new lemmas which are just weird permutations of the original ones. I've tried to name them with OD as a suffix so I don't get confused with the various primed lemmas that already existed.