fix(1000.yaml): correct reference to Bézout's identity
There are (at least) two results commonly attributed to Bézout: one is
Bézout's identity (a number theory lemma, which is in mathlib),
the other is Bézout's theorem (an algebraic geometry results, which apparently
is not).
To make matter worse, the 100 theorem project refers to the identity as
"Bézout's theorem" (which is arguably incorrect), while the 1000+ project
uses Bézout's theorem for the algebraic geometry result.

Correct 1000.yaml to account for this: fix the entry for Bézout's theorem,
and add the wikidata tag for Bézout's identity (with the mathlib declaration).
The latter tag is present on the wikipedia list, but not the 1000+ theorems
repository yet.