feat: let `notation3` handle `_` variables in extended binders
This PR enables writing terms like `⋃ _ ∈ s, Set.univ`, which is allowed by the syntax but has been so far unsupported. To get this to work, when the extended binder is expanded we need to replace `_` with a fresh identifier, since we need to refer to the variable again in the predicate.