feat(Data/List/Triwise): `List.Triwise`
Add a definition `List.Triwise`, for a predicate on a list determined
by a predicate on triples of distinct elements.

This is intended in particular for geometrical properties such as
betweenness, and was originally suggsted by @YaelDillies in 2021:
https://leanprover.zulipchat.com/#narrow/channel/116395-maths/topic/circular.20predicate
(definition adapated from there for Lean 4, and some API added).

There is undoubtedly lots more API that could be added, but what I
have here is sufficient for at least one geometrical use case:
defining a list of points being weakly in order on a line and proving
that this definition is equivalent to the list of points being the
image under a `lineMap` of an ordered list of values in the underlying
field.

Co-authored-by: Yaël Dillies <yael.dillies@gmail.com>