feat(CategoryTheory): the small object argument (#20245)
This concludes a series of PR towards the small object argument. In future PRs, it shall be used in order to formalise the proof by Grothendieck that Grothendieck abelian categories have enough injectives (#20079). It is also an important tool in Quillen's homotopical algebra, and it shall be used in the formalization of model category structures in homotopy theory and homological algebra.

In this PR, we introduce a typeclass `HasSmallObjectArgument I` which asserts that `I : MorphismProperty C` permits the small object argument. Under this assumption, we obtain `HasFunctorialFactorization I.rlp.llp I.rlp` and show that morphisms in `I.rlp.llp` are exactly the retracts of transfinite compositions of pushouts of coproducts of morphisms in `I`.

(Note: the small object argument was also formalised in Lean 3 in the pioneering work by Reid Barton by 2018.)



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>