chore(Algebra/Group): import cleanup for `Units/Defs.lean`
I noticed that this file imports `Logic.Function.Basic`, which is not a heavy import in absolute terms, but `Units.Defs` has been minimized quite a bit, so it would actually matter a bit in relative terms. Do we want to get rid of the import in exchange for doing one proof by hand?