feat: a module-finite R-algebra with the module topology is a topological ring (#21225)
This is the culmination of several PRs building `Mathlib/Topology/Algebra/Module/ModuleTopology.lean`, the API for the "module topology", a canonical topology for a module over a topological ring. The main result of this final PR is the fact that if R is a topological ring, if A is an R-algebra which is finite as an R-module, and if we give A the module topology, then A becomes a topological ring. The hard part, that multiplication is automatically continuous, is proved in this PR.

As an example of an application: the usual quaternions over the reals with the product topology coming from ℍ = ℝ⁴ are automatically a topological ring, because the product topology on a finite-dimensional vector space over a topological field is the module topology (another result in this file, but not in this PR). In particular you get the proof "for free" without having to argue explicitly that quaternion multiplication can be built via addition and multiplication of reals, which are continuous.

As far as I know this is original work, mostly due to Will Sawin [here](https://mathoverflow.net/a/477763/1384) (although I guess the argument in this PR is due to me). As in the other PRs I give informal proofs of everything in comments. 



Co-authored-by: Riccardo Brasca <riccardo.brasca@gmail.com>