chore: deprime last `cases'` for `Or`, etc. (#21945)
The main part of this PR is a follow-up to #16526. It removes (manually) all remaining uses of `cases'` used to split an `Or`, found in the output of [this run](https://github.com/leanprover-community/mathlib4/actions/runs/13352300428/job/37290385451).

We also remove (again manually) all `cases'` in `Data.QPF.Univariate.Basic` and `Order.Interval.Finset.Defs`, which were generating multiline outputs in the run.