chore(Algebra/Order/Group): downstream lemmas in `Defs.lean`
This PR moves a few lemmas in `Algebra.Order.Group.Defs` downstream to `Order.Group.Basic`, so that we can define ordered group with minimal imports. It means we have to add some extra imports of `Order.Group.Basic` to other files, so let's see how much we actually save with this change.