chore(Data/Int): split `Defs` into `Init` and `Basic`
This PR follows the instructions in a TODO comment in `Data/Int/Defs.lean` to split it into an `Init.lean` file for everything that doesn't meaningfully depend on Mathlib, and a `Basic.lean` file for things that do. I redid a few proofs in `Int.Init` so they are provable without reference to Mathlib.