feat: let `notation3` handle `_` variables in extended binders (#21924)
This PR enables writing terms like `⋃ _ ∈ s, Set.univ`, which is allowed by the syntax (and which is accepted by other notation that uses `extendedBinders` like `Exists`) but has been so far unsupported in `notation3`. To get this to work, when the extended binder is expanded we need to replace `_` with a fresh identifier, since we need to refer to the variable again in the predicate.