feat(CategoryTheory): commutation of coyoneda.obj to some colimits in Grothendieck abelian categories (#20014)
Let `C` be a Grothendieck abelian category. Let `X : C`. Assume that `κ` is a regular cardinal such that `Subobject X` is of cardinality `< κ`. Let `Y : J ⥤ C` be a functor from a `κ`-filtered category. We consider the map `colim_j (X  ⟶ Y_j) → (X ⟶ colim Y)`. We show that it is injective, and under the additional assumption that all maps `Y_j ⟶ Y_j'` are monomorphisms, it is bijective.



Co-authored-by: Joël Riou <37772949+joelriou@users.noreply.github.com>