chore(Data/Nat): split `Defs` into `Init` and `Basic` (#22008)
This PR follows the instructions in a TODO comment in `Data/Nat/Defs.lean` to split it into an `Init.lean` file for everything that doesn't meaningfully depend on Mathlib, and a `Basic.lean` file for things that do. I redid a few proofs in `Nat.Init` so they are provable without reference to Mathlib.