feat(CategoryTheory): characterization of injective objects in terms of lifting properties (#22104)
This PR adds the lemma `injective_iff_rlp_monomorphisms_zero` which shows that `I` is an injective object iff `0 : I ⟶ 0` has the right lifting property with respect to the class of monomorphisms. We also move the files regarding injectives objects to a subdirectory. (Dual results for projective objects are also added.)